Allow each thread in Map to process multiple elems
This modifies the behaviour of the default Tout(Tin) Map kernel so that
each work item can process multiple elements of the dataset, allowing an
N to 1 relationship between dataset elements and threads.

At the moment, the number of partitions to use is hardcoded. In the
future, this value could be set optimally at runtime.