Stencil: Combine padding types into single kernel
This replaces the three separate kernels for Neutral, Nearest, and
NearestInitial padding types into a single kernel, using preprocessor
defines to perform the correct border region loading behaviour.

While more concise than before, this code is still a way off of being
readable, so further work can yet be done.

NOTE that the "nearest" padding type is broken. It seems to be
performing out of bounds memory accesses leading to garbage values in
the output. This patch does not fix it, but has been tested with the
neutral and nearest_initial padding types and performs the same.